refactored classes into subdirectories
