+++ index.html
@@ -7,10 +7,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <Link rel="stylesheet" href="index.css">
     <title>Chess</title>
+
 </head>
 
 <body>
-    
+
     <div class="container">
         <h1 class="title">Chess!</h1>
         <h2 class="turn">White's turn</h2>
@@ -24,8 +25,9 @@
         </div>
         <div class="board"></div>
     </div>
-    
 
-    <script src="index.js"></script>
+
+    <script src="src/index.js" type="module" crossorigin></script>
 </body>
+
 </html>
