+++ package.json
@@ -10,6 +10,7 @@
   "author": "Best buds",
   "license": "MIT",
   "devDependencies": {
+    "jest": "^29.5.0",
     "nodemon": "^2.0.22"
   }
 }
