+++ src/move.js
@@ -1,0 +1,22 @@
+export default class Move {
+  static fromSquare(square, initiatingPiece, capturedPiece = null) {
+    return new Move(...square, initiatingPiece, capturedPiece);
+  }
+
+  constructor(row, file, initiatingPiece, capturedPiece) {
+    this.row = row;
+    this.file = file;
+    this.initiatingPiece = initiatingPiece;
+    this.sourceRow = initiatingPiece.row;
+    this.sourceFile = initiatingPiece.file;
+    this.capturedPiece = capturedPiece;
+    this.isCheck = false;
+  }
+
+  get isPawnDoubleMove() {
+    return (
+      this.initiatingPiece.name === "Pawn" &&
+      Math.abs(this.sourceRow - this.row) === 2
+    );
+  }
+}
