+++ src/pieces/index.mjs
@@ -1,0 +1,7 @@
+export { default as Piece } from "./piece.mjs"
+export { default as Pawn } from "./pawn.mjs"
+export { default as Rook } from "./rook.mjs"
+export { default as Knight } from "./knight.mjs"
+export { default as Bishop } from "./bishop.mjs"
+export { default as Queen } from "./queen.mjs"
+export { default as King } from "./king.mjs"
